+++ vote/src/App.tsx
@@ -4,18 +4,20 @@ import { RouterProvider } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import GlobalStyle from "@styles/globalStyle";
 import theme from "@styles/theme";
-import { ErrorHandler } from "@pages/ErrorHandler";
+import { QueryClient, QueryClientProvider } from "react-query";
+
+const queryClient = new QueryClient();
 
 function App() {
   return (
-    <Suspense fallback="..loading">
-      <ErrorHandler>
+    <QueryClientProvider client={queryClient}>
+      <Suspense fallback="..loading">
         <ThemeProvider theme={theme}>
           <RouterProvider router={Router} />
           <GlobalStyle />
         </ThemeProvider>
-      </ErrorHandler>
-    </Suspense>
+      </Suspense>
+    </QueryClientProvider>
   );
 }
