+++ vote/src/api/customAxios.ts
@@ -1,6 +1,16 @@
-import axios from "axios";
+// import axios from "axios";
 
-export const customAxios = axios.create({
+// export const customAxios = axios.create({
+//   baseURL: `${import.meta.env.VITE_APP_BASE_URL}`,
+//   headers: {
+//     "Content-Type": "application/json",
+//     "Access-Control-Allow-Origin": "*",
+//   },
+// });
+
+import axios, { AxiosInstance } from "axios";
+
+export const customAxios: AxiosInstance = axios.create({
   baseURL: `${import.meta.env.VITE_APP_BASE_URL}`,
   headers: {
     "Content-Type": "application/json",
