--- vote/src/components/layout/RootLayout.tsx
@@ -1,10 +1,0 @@
-import { Outlet } from "react-router-dom";
-
-/** RootLayout */
-export default function RootLayout() {
-  return (
-    <>
-      <Outlet />
-    </>
-  );
-}
